[examples] Re-organize CV examaples
