+++ examples/docs/action_recognition.md
@@ -4,7 +4,7 @@ Action recognition is a computer vision technique to infer human actions (presen
 
 In this example, you learn how to implement inference code with a [ModelZoo model](../../docs/model-zoo.md) to detect human actions in an image.
 
-The source code can be found at [ActionRecognition.java](https://github.com/deepjavalibrary/djl/blob/master/examples/src/main/java/ai/djl/examples/inference/ActionRecognition.java).
+The source code can be found at [ActionRecognition.java](https://github.com/deepjavalibrary/djl/blob/master/examples/src/main/java/ai/djl/examples/inference/cv/ActionRecognition.java).
 
 ## Setup Guide
 
@@ -22,7 +22,7 @@ Use the following command to run the project:
 
 ```sh
 cd examples
-./gradlew run -Dmain=ai.djl.examples.inference.ActionRecognition
+./gradlew run -Dmain=ai.djl.examples.inference.cv.ActionRecognition
 ```
 
 Your output should look like the following:
