+++ examples/src/main/java/ai/djl/examples/inference/cv/ObjectDetection.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"). You may not use this file except in compliance
  * with the License. A copy of the License is located at
@@ -10,14 +10,15 @@
  * OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions
  * and limitations under the License.
  */
-package ai.djl.examples.inference;
+package ai.djl.examples.inference.cv;
 
+import ai.djl.Application;
 import ai.djl.ModelException;
+import ai.djl.engine.Engine;
 import ai.djl.inference.Predictor;
 import ai.djl.modality.cv.Image;
 import ai.djl.modality.cv.ImageFactory;
 import ai.djl.modality.cv.output.DetectedObjects;
-import ai.djl.modality.cv.translator.YoloV5TranslatorFactory;
 import ai.djl.repository.zoo.Criteria;
 import ai.djl.repository.zoo.ZooModel;
 import ai.djl.training.util.ProgressBar;
@@ -35,56 +36,56 @@ import java.nio.file.Paths;
  * An example of inference using an object detection model.
  *
  * <p>See this <a
- * href="https://github.com/deepjavalibrary/djl/blob/master/examples/docs/mask_detection.md">doc</a>
+ * href="https://github.com/deepjavalibrary/djl/blob/master/examples/docs/object_detection.md">doc</a>
  * for information about this example.
  */
-public final class MaskDetection {
+public final class ObjectDetection {
 
-    private static final Logger logger = LoggerFactory.getLogger(MaskDetection.class);
+    private static final Logger logger = LoggerFactory.getLogger(ObjectDetection.class);
 
-    private MaskDetection() {}
+    private ObjectDetection() {}
 
     public static void main(String[] args) throws IOException, ModelException, TranslateException {
-        DetectedObjects detection = MaskDetection.predict();
+        DetectedObjects detection = predict();
         logger.info("{}", detection);
     }
 
     public static DetectedObjects predict() throws IOException, ModelException, TranslateException {
-        // To feed in local image, use ImageFactory.getInstance().fromFile(...)
-        String imageUrl = "https://resources.djl.ai/images/face_mask_detection/face_mask.png";
-        Image img = ImageFactory.getInstance().fromUrl(imageUrl);
+        Path imageFile = Paths.get("src/test/resources/dog_bike_car.jpg");
+        Image img = ImageFactory.getInstance().fromFile(imageFile);
 
-        // modelUrl can be replaced to local onnx model file
-        String modelUrl = "https://resources.djl.ai/demo/onnxruntime/face_mask_detection.zip";
+        String backbone;
+        if ("TensorFlow".equals(Engine.getDefaultEngineName())) {
+            backbone = "mobilenet_v2";
+        } else {
+            backbone = "resnet50";
+        }
+
         Criteria<Image, DetectedObjects> criteria =
                 Criteria.builder()
+                        .optApplication(Application.CV.OBJECT_DETECTION)
                         .setTypes(Image.class, DetectedObjects.class)
-                        .optModelUrls(modelUrl)
-                        .optEngine("OnnxRuntime")
-                        .optTranslatorFactory(new YoloV5TranslatorFactory())
+                        .optFilter("backbone", backbone)
+                        .optEngine(Engine.getDefaultEngineName())
                         .optProgress(new ProgressBar())
-                        .optArgument("optApplyRatio", true) // post process
-                        .optArgument("rescale", true) // post process
                         .build();
 
-        try (ZooModel<Image, DetectedObjects> model = criteria.loadModel()) {
-            try (Predictor<Image, DetectedObjects> predictor = model.newPredictor()) {
-                DetectedObjects detection = predictor.predict(img);
-                String outputDir = "build/output";
-                saveBoundingBoxImage(img, detection, outputDir);
-                return detection;
-            }
+        try (ZooModel<Image, DetectedObjects> model = criteria.loadModel();
+                Predictor<Image, DetectedObjects> predictor = model.newPredictor()) {
+            DetectedObjects detection = predictor.predict(img);
+            saveBoundingBoxImage(img, detection);
+            return detection;
         }
     }
 
-    private static void saveBoundingBoxImage(Image img, DetectedObjects detection, String outputDir)
+    private static void saveBoundingBoxImage(Image img, DetectedObjects detection)
             throws IOException {
-        Path outputPath = Paths.get(outputDir);
-        Files.createDirectories(outputPath);
+        Path outputDir = Paths.get("build/output");
+        Files.createDirectories(outputDir);
 
         img.drawBoundingBoxes(detection);
 
-        Path imagePath = outputPath.resolve("face_mask_result.png");
+        Path imagePath = outputDir.resolve("detected-dog_bike_car.png");
         // OpenJDK can't save jpg with alpha channel
         img.save(Files.newOutputStream(imagePath), "png");
         logger.info("Detected objects image has been saved in: {}", imagePath);
